Inline single-use extension helpers (#431)
